- add docker file - use common dto
