+++ common/README.md
@@ -1,0 +1,14 @@
+# Common
+gradle clean build
+
+* Common Dto
+mvn install:install-file -Dfile=/Users/nguyendangtuan/Documents/Project/microservice/common/common-dto/build/libs/common-dto-0.0.1-SNAPSHOT.jar -DgroupId=com.dangtuan -DartifactId=common-dto -Dversion=0.0.1-SNAPSHOT -Dpackaging=jar
+
+* Common Entity
+mvn install:install-file -Dfile=/Users/nguyendangtuan/Documents/Project/microservice/common/common-entity/build/libs/common-entity-0.0.1-SNAPSHOT.jar -DgroupId=com.dangtuan -DartifactId=common-entity -Dversion=0.0.1-SNAPSHOT -Dpackaging=jar
+
+* Common Resource
+mvn install:install-file -Dfile=/Users/nguyendangtuan/Documents/Project/microservice/common/common-resource/build/libs/common-resource-0.0.1-SNAPSHOT.jar -DgroupId=com.dangtuan -DartifactId=common-resource -Dversion=0.0.1-SNAPSHOT -Dpackaging=jar
+
+* Common Util
+mvn install:install-file -Dfile=/Users/nguyendangtuan/Documents/Project/microservice/common/common-util/build/libs/common-util-0.0.1-SNAPSHOT.jar -DgroupId=com.dangtuan -DartifactId=common-util -Dversion=0.0.1-SNAPSHOT -Dpackaging=jar
