+++ common/build.gradle
@@ -5,7 +5,7 @@ allprojects {
 
     repositories {
         mavenCentral()
-	    mavenLocal()
+        mavenLocal()
     }
 }
