+++ common/common-dto/build.gradle
@@ -4,13 +4,13 @@ version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '1.8'
 
 dependencies {
-	annotationProcessor 'org.projectlombok:lombok:1.18.12'
-	compile 'com.fasterxml.jackson.core:jackson-databind:2.10.1'
-	compile 'org.apache.commons:commons-lang3:3.10'
-	compile 'org.apache.logging.log4j:log4j-to-slf4j:2.13.3'
-	compile 'org.hibernate:hibernate-core:5.4.18.Final'
-	compile 'org.hibernate.validator:hibernate-validator:6.1.5.Final'
-	compile 'org.projectlombok:lombok:1.18.12'
-	compile 'org.springframework:spring-tx:5.2.8.RELEASE'
-	compile 'org.springframework:spring-web:5.2.8.RELEASE'
+    annotationProcessor 'org.projectlombok:lombok:1.18.12'
+    compile 'com.fasterxml.jackson.core:jackson-databind:2.10.1'
+    compile 'org.apache.commons:commons-lang3:3.10'
+    compile 'org.apache.logging.log4j:log4j-to-slf4j:2.13.3'
+    compile 'org.hibernate:hibernate-core:5.4.18.Final'
+    compile 'org.hibernate.validator:hibernate-validator:6.1.5.Final'
+    compile 'org.projectlombok:lombok:1.18.12'
+    compile 'org.springframework:spring-tx:5.2.8.RELEASE'
+    compile 'org.springframework:spring-web:5.2.8.RELEASE'
 }
