+++ common/common-dto/src/main/java/com/dangtuan/dto/order/OrderDto.java
@@ -1,0 +1,32 @@
+package com.dangtuan.dto.order;
+
+import java.math.BigDecimal;
+import java.util.Date;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderDto {
+
+  private Long id;
+
+  private Long amount;
+
+  private BigDecimal price;
+
+  private String tenantId;
+
+  private Boolean deleted;
+
+  private Date createdDate;
+
+  private Date updatedDate;
+
+  private String createdBy;
+
+  private String updatedBY;
+
+}
