+++ common/common-resource/build.gradle
@@ -1,4 +1,3 @@
-archivesBaseName = 'common'
 group = 'com.dangtuan'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '1.8'
