+++ common/common-util/build.gradle
@@ -4,11 +4,11 @@ sourceCompatibility = '1.8'
 
 
 dependencies {
-	annotationProcessor 'org.projectlombok:lombok:1.18.12'
-	implementation('org.apache.commons:commons-lang3:3.10')
-	implementation 'org.apache.logging.log4j:log4j-to-slf4j:2.13.3'
-	implementation 'org.hibernate:hibernate-core:5.4.18.Final'
-	implementation 'org.hibernate.validator:hibernate-validator:6.1.5.Final'
-	implementation('org.projectlombok:lombok:1.18.12')
-	testImplementation 'org.testng:testng:6.9.10'
+    annotationProcessor 'org.projectlombok:lombok:1.18.12'
+    implementation('org.apache.commons:commons-lang3:3.10')
+    implementation 'org.apache.logging.log4j:log4j-to-slf4j:2.13.3'
+    implementation 'org.hibernate:hibernate-core:5.4.18.Final'
+    implementation 'org.hibernate.validator:hibernate-validator:6.1.5.Final'
+    implementation('org.projectlombok:lombok:1.18.12')
+    testImplementation 'org.testng:testng:6.9.10'
 }
