+++ config/src/main/resources/config/local/auth-server-local.yml
@@ -18,7 +18,7 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/oauth_server_db?serverTimezone=UTC
     username: root
-    password:
+    password: 123456
   jpa:
     show-sql: true
     properties.hibernate.dialect: org.hibernate.dialect.MySQL5InnoDBDialect
