+++ config/src/main/resources/config/local/order-service-local.yml
@@ -17,7 +17,7 @@ spring:
   datasource:
     url: jdbc:mysql://localhost:3306/order_db?serverTimezone=UTC
     username: root
-    password:
+    password: 123456
   jpa:
     properties.hibernate.dialect: org.hibernate.dialect.MySQL5InnoDBDialect
     hibernate:
