+++ config/src/main/resources/config/local/resource-server-local.yml
@@ -18,7 +18,7 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/resource_server_db?serverTimezone=UTC
     username: root
-    password:
+    password: 123456
   jpa:
     properties.hibernate.dialect: org.hibernate.dialect.MySQL5InnoDBDialect
     hibernate:
