+++ order-service/build.gradle
@@ -1,8 +1,21 @@
+buildscript {
+    repositories {
+        maven {
+            url "https://plugins.gradle.org/m2/"
+        }
+    }
+    dependencies {
+        classpath "com.palantir.gradle.docker:gradle-docker:0.25.0"
+    }
+}
+
 plugins {
     id 'org.springframework.boot' version '2.3.2.RELEASE'
     id 'io.spring.dependency-management' version '1.0.9.RELEASE'
     id 'java'
 }
+
+apply plugin: 'com.palantir.docker'
 
 group = 'com.dangtuan'
 version = '0.0.1-SNAPSHOT'
@@ -24,7 +37,7 @@ ext {
 dependencies {
     annotationProcessor 'org.mapstruct:mapstruct-processor:1.3.1.Final'
     implementation 'com.auth0:java-jwt:3.8.2'
-//    implementation 'com.dangtuan:common-dto:0.0.1-SNAPSHOT changing = true'
+    implementation 'com.dangtuan:common-dto:0.0.1-SNAPSHOT'
     implementation 'org.mapstruct:mapstruct:1.3.1.Final'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -52,4 +65,16 @@ dependencyManagement {
 }
 test {
     useJUnitPlatform()
+}
+
+task unpack(type: Copy) {
+    dependsOn bootJar
+    from 'build/libs'
+    into("build/dependency")
+}
+
+docker {
+    name "${project.group}/${bootJar.baseName}"
+    copySpec.from(tasks.unpack.outputs).into("dependency")
+    buildArgs(['DEPENDENCY': "dependency"])
 }
